+++ wetterdienst/provider/dwd/observation/download.py
@@ -5,7 +5,6 @@
 from __future__ import annotations
 
 import logging
-from concurrent.futures import ThreadPoolExecutor
 from io import BytesIO
 from typing import TYPE_CHECKING
 from zipfile import BadZipFile
@@ -14,7 +13,7 @@ from fsspec.implementations.zip import ZipFileSystem
 
 from wetterdienst.exceptions import ProductFileNotFoundError
 from wetterdienst.metadata.cache import CacheExpiry
-from wetterdienst.util.network import download_file
+from wetterdienst.util.network import File, download_files
 
 if TYPE_CHECKING:
     import polars as pl
@@ -27,38 +26,35 @@ log = logging.getLogger(__name__)
 def download_climate_observations_data(
     remote_files: pl.Series,
     settings: Settings,
-) -> list[tuple[str, BytesIO]]:
+) -> list[File]:
     """Download climate observations data from DWD server."""
-    if len(remote_files) > 1:
-        with ThreadPoolExecutor() as p:
-            files_in_bytes = p.map(
-                lambda file: _download_climate_observations_data(remote_file=file, settings=settings),
-                remote_files,
-            )
-    else:
-        files_in_bytes = [_download_climate_observations_data(remote_file=remote_files[0], settings=settings)]
-    return list(zip(remote_files, files_in_bytes, strict=False))
-
-
-def _download_climate_observations_data(remote_file: str, settings: Settings) -> BytesIO:
-    return BytesIO(__download_climate_observations_data(remote_file=remote_file, settings=settings))
-
-
-def __download_climate_observations_data(remote_file: str, settings: Settings) -> bytes:
-    file = download_file(
-        url=remote_file,
+    files = download_files(
+        urls=remote_files.to_list(),
         cache_dir=settings.cache_dir,
         ttl=CacheExpiry.FIVE_MINUTES,
         client_kwargs=settings.fsspec_client_kwargs,
         cache_disable=settings.cache_disable,
     )
+    # filter out exceptions
+    files = [file for file in files if isinstance(file.content, BytesIO)]
+    # unpack the files
+    return [_unpack_climate_observations_data(file) for file in files]
+
+
+def _unpack_climate_observations_data(file: File) -> File:
     try:
-        zfs = ZipFileSystem(file)
+        zfs = ZipFileSystem(file.content)
     except BadZipFile as e:
-        msg = f"The archive of {remote_file} seems to be corrupted."
+        msg = f"The archive of {file.filename} seems to be corrupted."
         raise BadZipFile(msg) from e
     product_file = zfs.glob("produkt*")
     if len(product_file) != 1:
-        msg = f"The archive of {remote_file} does not hold a 'produkt' file."
+        msg = f"The archive of {file.filename} does not hold a 'produkt' file."
         raise ProductFileNotFoundError(msg)
-    return zfs.open(product_file[0]).read()
+    product_bytes = zfs.open(product_file[0]).read()
+    # let's put this again in another file object
+    return File(
+        url=file.url,
+        content=BytesIO(product_bytes),
+        status=file.status,
+    )