+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Make retry of `download_file` more robust
 - Overhaul docs switching to `sphinx` and `myst-parser`
 - Improve exception handling in restapi
+- Improve download of files
 
 ### Fix
 - Export: Fix influx tags and fields